updates to report graphs for banking

+++ src/app/facility/analysis/run-analysis/facility-analysis/facility-analysis.component.ts
@@ -45,7 +45,7 @@ export class FacilityAnalysisComponent implements OnInit {
     let facilityMeterData: Array<IdbUtilityMeterData> = this.utilityMeterDataDbService.facilityMeterData.getValue();
     let accountPredictorEntries: Array<IdbPredictorData> = this.predictorDataDbService.accountPredictorData.getValue();
     let accountPredictors: Array<IdbPredictor> = this.predictorDbService.accountPredictors.getValue();
-    if (typeof Worker !== 'undefined') {
+    if (typeof Worker !== 'undefined' && false) {
       this.worker = new Worker(new URL('src/app/web-workers/annual-facility-analysis.worker', import.meta.url));
       this.worker.onmessage = ({ data }) => {
         this.worker.terminate();
@@ -78,6 +78,7 @@ export class FacilityAnalysisComponent implements OnInit {
       let monthlyAnalysisSummaryData: Array<MonthlyAnalysisSummaryData> = annualAnalysisSummaryClass.monthlyAnalysisSummaryData;
       this.analysisService.annualAnalysisSummary.next(annualAnalysisSummaries);
       this.analysisService.monthlyAccountAnalysisData.next(monthlyAnalysisSummaryData);
+      this.analysisService.calculating.next(false);
     }
   }
